DFA-312 Registration Form
Port registration form from FreeMarker to Nunjucks
Port validation from Java to TypeScript

I'm not in love with the layout or structure of this and the validation stuff could definitely do with a re-factoring but hey, it works and it will make more sense to refactor when we have a new form.

Next job - do the Sheets stuff with our definitely valid data. And some tests probably.  That would be a good idea.