DFA-312 Persist form data
We need to save form data.  The methods we are using are saving to Goolge Sheets and saving the data to an S3 bucket in case anything happens to the Sheet.

Previously this was done using a Java application so for local frontend development it was sufficient to just run the Ruby/Minuteman application.  Now the application is being rewritten in Express which means that we need a way of running it locally without persisting the data because it's a bit much having to set up an AWS account, creating and S3 bucket, sorting out a service account for Google Sheets and a dummy sheet with the right headings when all you want to do is tweak a stylesheet.

Unfortunately, I'm not familiar with how conditional dependencies work in JavaScript so I've used a Strategy/Delegation pattern to do this as best I could.  At this point I relase that this scenario was the answer to "why would we want lambdas when we could do it all in JavaScript".  It is what it is and I refuse to litte code with if statements when they can be hidden away inside a 'thing' and the 'thing' decides what to implement.

Anyway, it is what it is. I'd much prefer being able to inject instances of classes instead of wrapping them but I don't know how that's done in JavaScript.

Add a proliferation of classes.
Add .env files and use them.
Update the README.md