Add extended email domains
Use Pay's list of legit gov emails in a text file.
Append this list to the short list we have already.
Use the `find` stream operation to shortcut all the .gov.uk addresses.

Ignore the fact that the file is re-read every time the form is submitted; once we've refactored our routes the list can be loaded into a single instance of the validator and injected.  Assuming that's fine with this event based model stuff.