Request Private Beta Form
Refactor validation to have checked values passed in.
Refactor Sheets to take a data Range for destination data.
Create `request.njk` view to render new form.
Add new routes for new view with necessary values passed in.
Add extra env vars to `.env.example`
Add a slash to the css link in the base template

Points to the success and failure pages for "Register" and these should be updated when the other pages become available.