DFA-151 Request submitted page
This adds the content as shown in the ticket.

Not sure about what to do with the links in the header part.

The spacing might be a bit off from a design point of view.  I'm not sure what the correct tweak is or how it should be dealt with.

On the plus side, the form can now point to something on success.