WIP
Still a work in progress.

I can see some areas for refactoring - the contact links should probably be done with a table so we enforce the same pattern over all pages.  Chances are this is redundant anyway.

Likewise the side menu should be checked on each page but should probably be treated as a feature in its own right.