DFA-376: Deploy a preview app for each pull request
Add a "Test PR" workflow that deploys a preview app each time a pull request is opened or updated.
The workflow uses reusable workflows which will be later used in the production pipeline as well.
The preview pipeline builds the app, runs unit tests, deploys to PaaS and runs acceptance tests (stub job for now).
Each app is deployed on a unique URL derived from the branch name.

  - Cache npm modules
  - Run unit tests using mocha