DFA-376: Handle empty config values
Some config values are not present in non-prod environments yet `cf push` requires all env vars in the manifest file to be preset when deploying. Fix this by using env vars for the optional vars and use shell parameter expansion to substitute a white space if a value is not present.