DFA-28 trim fields before validating
Trim form fields before validating so that a whitespace field is considered to be blank.

This is implemented so that the whitespace is permanently removed, not just as part of the validation.

Also fix type error in ZendeskService which surfaced once the values were being trimmed and passed in.

Remove S3 code as it is no longer needed but would need updaating to use the new Map<string, string> representation of the form input.

Modify sheets related code to take Map<string, string> so that trimmed input is used.