DFA-347 Write acceptance test
Write a suite of acceptance tests for the product.

Update `package.json` with some commands to run features and necessary packages.
Create directory structure for running tests with cucumber-js.
Write a number of features for various pages using ACs in JIRA tickets (modified where necessary to make the testing work).
Write steps to execute scenarios and world-class to provide puppeteer.

There are some pages that aren't included here.  I couldn't find the JIRA tickets for them.  We should look at including them all.

There should be some accessibility specific features.  We'd need to figure a lot out but we should definitely have a failure if the accessibility statement is out of date.  There are other things like making sure every page has a working main content link.

The contact scenarios are repeated quite a lot.  It might be worth having a feautre that runs the same steps on each page using a scenario outline and a table of pages.  This will mean it's easier to update if and when the email address goes.