DFA-478 Fix slow running tests
Refactor contact links in the decide/evaluate pages.  They've been moved to a single feature file.  This means that if the contact methods are altered there will only be one file / step to alter.

Also put the code used by this steps in a different file because they'll be used later in this PR.  The IDE doesn't like it but it seems to build and run fine.