Update tests that follow links to external sites
Add a shared step so that we can check whether a page contains a link and whether the link points to the correct URL
Use that step in the various scenarios marked with @wip
Remove @wip tags
Amend a URL in the `decide-user-journeys.njk` because it had a stray backslash at the end.

Remove @wip from the end user scenario in the support feature as we can't test it any other way.  It should be fine though right?