BAU Add steps for Zendesk and Sheets forms
Add a step to check the content of the success page for forms that post to Zendesk or Sheets.

We still need a feature file for the register pages but we're waiting for the ACs.

Comment out `rootDir` in `tsconfig.json`; typescript can still find what it needs but IntelliJ no longer gets upset when you include the shared functions in a step.