DFA-551 Smoke tests
Add monitoring.

Entire canary stack can be deployed via SAM.  We only have one simple lambda so it's written inline in the template.  If we don't do that, we need to create a bucket, build and upload canaries and stuff.
Create `smoke-tests.yml` for a GitHub action that will deploy smoke tests if anything in `monitoring/` changes.
Create `github-actions-oidc-provider-template.yaml` so an OIDC provider can be created for a given AWS account.