Two SAM configurations - one with a lambda that calls a lambda and another that gets called by a lambda.
Confused? It is confusing.  The real meat and bones is with the Roles and Policies (Trust Policy and Permissions Policies - because they're two different things).

You could use sam to build and deploy these to new stacks but you'd have to dig out the various outputs and paste them into the templates.  You'll see one has some parameters that aren't used.  They hark back to a time when I niaively thought I could just write a template that would work.  Such happy times before the dark clouds of knowledge and experience beat me down into the click-ops and copy pattern.  Anyway, if you go into the integration account and invoke the CallingFunction lambda (with a test paylod of {"message": "whatever message you want"}) you'll see a response from the lambda in the development account.

There's more work to do figuring out the best way to pass things in.  I initially used ENV_VARS for the lambdas and I think that's still the best way.  Cross-Account exporting isn't a thing. It seems you can have a lambda that returns a list of exports in the account but you still need to do the bootstrapping of roles and stuff that these two stacks do.  It is a way we could avoid too much clickery in future.