BAU: Refactor and fix smoke tests workflow
Use the shared SAM actions to build and deploy the test stack

Don't pass the template argument to the sam deploy command as that causes the CLI to not upload the node dependencies and hence the function doesn't work.

Better result and error handling for the snsToSlack lambda function. Return a promise or throw an error to let AWS handle the result according to the error handling settings of the function. Returning a promise lets AWS ensure it gets resolved. See https://docs.aws.amazon.com/lambda/latest/dg/nodejs-handler.html#nodejs-handler-async
  - Throw errors when necessary
  - Return a promise from the async handler