BAU: Fix and refactor smoke tests
Convert the Canary SNS to Slack notification script into TypeScript
  - Build the script using esbuild
  - Import and use types for static type checking

Add additional parameters to the template
Resolve the parameter values dynamically when deploying the template. This removes the need to get the parameter value using the SSM API in the lambda function.

Add additional checks to the Canary script
Check the response status is 2xx, otherwise fail the step.
By default, Synthetics treats the step as passing when it manages to retrieve a page even if the status code is not 2xx. The explicit check is needed to ensure the step will fail if the product page can't be successfully accessed.

Use a dead-letter queue
Create an SNS topic to serve as a DLQ for the lambda function. By default, the lambda function will retry delivering failing messages two times. After the third attempt the message will be posted to the SNS topic.
The SNS topic has an email subscription to the DfA tech group. All messages that fail to notify to Slack will be emailed to that group so failures can be investigated and rectified.

Use the correct Canary name for the CloudWatch Alarm metric

Use blocks and attachments when posting a message to Slack to display a richer message with a coloured bar indicating the status of the product pages.