BAU: Fix bug in email validator
Check that the whole part after the @ sign in the email matches one of the domains to make sure domains that end in a valid domain but prefixed with an arbitrary string are rejected. This means that domains such as foo.gov.uk are now rejected where they were previously accepted.