SSE-3210 Create test constraints for Reg Interest Jira Form
We are going to add a Jira integration to this repository so we can
connect our external register your interest form to the internal
Onboarding Team Practices (which are heavily Jira based). This is both
something new, but also a swap out for similar existing processes and
code that sits in this repo.

We currenty POST the result to a google sheet. Jira is the replacement
for that, though the two will live side by side for a bit.

Star that process by creating some test data that represents both the
kinds of submissions we get off the form, and also some of the key data
we'd need to populate a Jira ticket

Co-authored-by: Ryan Andrews <ryan.andrews@digital.cabinet-office.gov.uk>
Co-authored-by: Huw Diprose <huw.diprose@digital.cabinet-office.gov.uk>