SSE-3210 Create stubJiraService
We're following the existing model here set out in the repo, we'll make
a stub class that we can call on selectively to stub out calls that
would have otherwise hit a real Jira instance.

Also shows the basic shape of what we're working towards with the real
service, a class that akes in the board URL and a project key, and has a
single function to POST the result, returning a small confirmation
payload to get logged.

Co-authored-by: Ryan Andrews <ryan.andrews@digital.cabinet-office.gov.uk>
Co-authored-by: Huw Diprose <huw.diprose@digital.cabinet-office.gov.uk>