SSE-2894 Remove mocha-sonarqube-reporter
This is a specific package that's designed to help you take coverage
output and put it into an XML format the sonarcloud apparently speaks.

https://www.npmjs.com/package/mocha-sonarqube-reporter

Interestingly I can't see anywhere else using this across the program:
https://github.com/search?q=org%3Agovuk-one-login%20mocha-sonarqube-reporter&type=code

In addition, seems like this was never configured correctly, the next
line after using this as the reporter should be a test output
directory, as we never gave that it's not actually doing anything

Our sonarr config seems to be using the lcov report anyway!
https://github.com/govuk-one-login/onboarding-product-page/blob/912791602e12aa6f63fc6daecb97ecf833073205/sonar-project.properties#L10

So on that basis, removing a dependency here and simplifying further