SSE-2894 Remove --exit from test:cov
mocha docs call out --exit as a solution to test hanging behaviour after
upgrading to v4.0.0+ of mocha.

https://mochajs.org/#-exit

> TL;DR: If your tests hang after an upgrade to Mocha v4.0.0 or newer,
use --exit for a quick (though not necessarily recommended) fix.

it goes on to say:

> The easiest way to “fix” the issue is to pass --exit to the Mocha
process. It can be time-consuming to debug — because it’s not always
obvious where the problem is — but it is recommended to do so.

As our tests are not hanging right now this line is redundant, we should
look to debug if we get a reoccurance