SSE-2894 Remove TS_NODE_COMPILER_OPTIONS from npm run test
So looking at what we're doing here, we're giving it an env target of
commonjs.

What concerns me is that if I check our tsconfig.json we're targeting
es2020. I'm concerned about us testing against a different target from
our usual build. Stripping this off seems to lead to no change in test
coverage, but I think the risks of keeping this are higher than removing
it.