SSE-2894 Configure nyc with an nycrc.json and expand coverage to --all
At the core of SSE-2894 is a mystery. Why should sonarcloud be reporting
a lower coverage level than running npm run test:cov turns up.

It's become apparent we've got two different means to calcualte coverage
here, but the impact has to be bigger than just a slight change of
calc.

At a close look it seemed SonarCloud was reporting on a wider range of
files than NYC was. NYC was only picking up on /lib, because we only
have tests for things in the /lib.

https://sonarcloud.io/component_measures?metric=coverage&view=list&id=di-onboarding-product-page

Sonarcloud config seems fine for the moment, the issue is with NYC
default configuration.

NYC does not seemt to report on the absence of test files, instead of it
giving you a zero for coverage by default it assumes no coverage was
needed. Sonarcloud does not operate under the same assumptions.

The solution is is to move nyc over in two ways:

1. Let's make the config more detailed and explicit with a .nycrc
2. Let's make one of those configurations nyc's --all flag

https://github.com/istanbuljs/nyc/blob/main/README.md#common-configuration-options

All will allow us to see the full scope of the app and missing coverage.
It does risk us being too broad, but we can now do exclude / include to
handle that, same as we do in the sonarcloud config